% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygCAPTCHAv3.R
\name{GreCAPTCHAv3Server}
\alias{GreCAPTCHAv3Server}
\title{GreCAPTCHAv3Server}
\usage{
GreCAPTCHAv3Server(secretKey, reCaptchaResponse)
}
\arguments{
\item{secretKey}{The Secret key generated in Google reCAPTCHA admin console for your domain}

\item{reCaptchaResponse}{The response received from GreCAPTCHAv3Ui 'Field'}
}
\description{
Server part of the package to receive response token from invoked UI script
}
