% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygCAPTCHAv3.R
\docType{package}
\name{shinygCAPTCHAv3}
\alias{shinygCAPTCHAv3}
\title{shinygCAPTCHAv3}
\description{
Easy implementation of Google reCAPTCHAv3 in RShiny Web Applications.
}
\details{
\code{shinygCAPTCHAv3} can be implemented as follows.

Add the function \code{GreCAPTCHAv3Ui} in the UI part of your RShiny web application to invoke reCAPTCHAv3 Javascript.

Add the function \code{GreCAPTCHAv3Server} in the server part of your RShiny web application to receive response token and verify it with Google's servers for verification.

Click \href{https://sarthi2395.shinyapps.io/shinygCAPTCHAv3}{here} to see a demo.
}
